+++ lib/samanage/api/custom_forms.rb
@@ -7,13 +7,13 @@ module Samanage
 		end
 
 		# Get all custom forms
-		def collect_custom_forms
+		def collect_custom_forms(options: {})
 			page = 1
 			custom_forms = Array.new
 			total_pages = self.get_custom_forms[:total_pages]
-			while page <= total_pages
+			1.upto(total_pages) do |page|
+				puts "Collecting Custom Forms page: #{page}/#{total_pages}" if options[:verbose]
 				custom_forms += self.execute(http_method: 'get', path: "custom_forms.json?page=#{page}")[:data]
-				page += 1
 			end
 			custom_forms
 		end