+++ lib/samanage/api/departments.rb
@@ -5,11 +5,12 @@ module Samanage
 			self.execute(path: url)
 		end
 
-		def collect_departments
+		def collect_departments(options: {})
 			page = 1
 			departments = Array.new
 			total_pages = self.get_departments[:total_pages]
-			while page <= total_pages
+			1.upto(total_pages) do |page|
+				puts "Collecting Groups page: #{page}/#{total_pages}" if options[:verbose]
 				departments += self.execute(http_method: 'get', path: "departments.json?page=#{page}")[:data]
 				page += 1
 			end
@@ -18,6 +19,12 @@ module Samanage
 
 		def create_department(payload: nil, options: {})
 			self.execute(path: PATHS[:department], http_method: 'post', payload: payload)
+
 		end
+		def delete_department(id: )
+      self.execute(path: "departments/#{id}.json", http_method: 'delete')
+    end
+
+		alias_method :departments, :collect_departments
 	end
 end