+++ lib/samanage/api/incidents.rb
@@ -8,14 +8,27 @@ module Samanage
 		end
 
 
-		# Returns all incidents
-		def collect_incidents
-			page = 1
+		# Returns all incidents. 
+		# Options: 
+		#   - audit_archives: true
+		#   - layout: 'long'
+		def collect_incidents(options: {})
 			incidents = Array.new
 			total_pages = self.get_incidents[:total_pages]
-			while page <= total_pages
-				incidents += self.execute(http_method: 'get', path: "incidents.json?page=#{page}")[:data]
-				page += 1
+			puts "Pulling Incidents with Audit Archives (this may take a while)" if options[:audit_archives] && options[:verbose]
+			1.upto(total_pages) do |page|
+				puts "Collecting Incidents page: #{page}/#{total_pages}" if options[:verbose]
+				if options[:audit_archives]
+					archives = 'layout=long&audit_archive=true'
+					paginated_incidents = self.execute(path: "incidents.json?page=#{page}")[:data]
+					paginated_incidents.map do |incident|
+						archive_path = "incidents/#{incident['id']}.json?#{archives}"
+						incidents << self.execute(path: archive_path)[:data]
+					end
+				else
+					layout = options[:layout] == 'long' ? '&layout=long' : nil
+					incidents += self.execute(http_method: 'get', path: "incidents.json?page=#{page}#{layout}")[:data]
+				end
 			end
 			incidents
 		end
@@ -37,5 +50,12 @@ module Samanage
 			path = "incidents/#{id}.json"
 			self.execute(path: path, http_method: 'put', payload: payload)
 		end
+
+		def delete_incident(id: )
+      self.execute(path: "incidents/#{id}.json", http_method: 'delete')
+    end
+
+
+	alias_method :incidents, :collect_incidents
 	end
 end