+++ lib/samanage/api/users.rb
@@ -1,7 +1,6 @@
 module Samanage
 	class Api
 
-
 	# Get users, using URL builder
 	def get_users(path: PATHS[:user], options: {})
     url = Samanage::UrlBuilder.new(path: path, options: options).url
@@ -9,13 +8,13 @@ module Samanage
 	end
 
   # Returns all users in the account
-		def collect_users
+		def collect_users(options: {})
 			page = 1
 			users = Array.new
 			total_pages = self.get_users[:total_pages]
-			while page <= total_pages
+			1.upto(total_pages) do |page|
+				puts "Collecting Users page: #{page}/#{total_pages}" if options[:verbose]
 				users += self.execute(http_method: 'get', path: "users.json?page=#{page}")[:data]
-				page += 1
 			end
 			users
 		end
@@ -34,20 +33,20 @@ module Samanage
 		# Email is unique so compare first for exact match only. Returns nil or the id
 		def find_user_id_by_email(email: )
 			api_call = self.check_user(value: email)
-			api_call.dig(:data).select{|u| u['email'].to_s.downcase == email.to_s.downcase}.to_h['id']
+			api_call[:data].select{|u| u['email'].to_s.downcase == email.to_s.downcase}.first.to_h['id']
 		end
 
 		# Returns nil if no matching group_id
 		def find_user_group_id_by_email(email: )
 			user = self.check_user(value: email)
-			group_ids = user[:data].select{|u| u['email'].to_s.downcase == email.to_s.downcase}.to_h['group_ids']
+			group_ids = user[:data].select{|u| u['email'].to_s.downcase == email.to_s.downcase}.first.to_h['group_ids'].to_a
 			group_ids.each do |group_id|
 				group = self.find_group(id: group_id)
 				if group[:data]['is_user'] && email == group[:data]['email']
 					return group_id
 				end
 			end
-			return
+			return nil
 		end
 
     # Check for user by field (ex: users.json?field=value)
@@ -64,5 +63,11 @@ module Samanage
 			path = "users/#{id}.json"
 			self.execute(path: path, http_method: 'put', payload: payload)
 		end
+
+		def delete_user(id: )
+      self.execute(path: "users/#{id}.json", http_method: 'delete')
+		end
+
+	alias_method :users, :collect_users
 	end
 end