+++ spec/api/samanage_custom_field_spec.rb
@@ -3,13 +3,13 @@ require 'samanage'
 describe Samanage::Api do
 	context 'Custom Field' do
 		describe 'API Functions' do
-			before(:each) do
+			before(:all) do
 				TOKEN ||= ENV['SAMANAGE_TEST_API_TOKEN']
-				@controller = Samanage::Api.new(token: TOKEN)
+				@samanage = Samanage::Api.new(token: TOKEN)
+				@custom_fields = @samanage.custom_fields
 			end
 			it 'collects all custom fields' do
-				api_call = @controller.collect_custom_fields
-				expect(api_call).to be_a(Array)
+				expect(@custom_fields).to be_a(Array)
 			end
 		end
 	end