+++ spec/api/samanage_site_spec.rb
@@ -1,20 +1,21 @@
 require 'samanage'
 describe Samanage::Api do
 	context 'Site' do
-		before(:each) do
+		before(:all) do
 			TOKEN ||= ENV['SAMANAGE_TEST_API_TOKEN']
-			@controller = Samanage::Api.new(token: TOKEN)
+			@samanage = Samanage::Api.new(token: TOKEN)
+			@sites = @samanage.sites
 		end
 		it 'get_users: it returns API call of users' do
-			api_call = @controller.get_sites
+			api_call = @samanage.get_sites
 			expect(api_call).to be_a(Hash)
 			expect(api_call[:total_count]).to be_an(Integer)
 			expect(api_call).to have_key(:response)
 			expect(api_call).to have_key(:code)
 		end
 		it 'collects all sites' do
-			sites = @controller.collect_sites
-			site_count = @controller.get_sites[:total_count]
+			sites = @sites
+			site_count = @samanage.get_sites[:total_count]
 			expect(sites).to be_an(Array)
 			expect(sites.size).to eq(site_count)
 		end
@@ -29,11 +30,21 @@ describe Samanage::Api do
 					description: site_description
 				}
 			}
-			site_create = @controller.create_site(payload: payload)
+			site_create = @samanage.create_site(payload: payload)
 
 			expect(site_create[:data]['id']).to be_an(Integer)
 			expect(site_create[:data]['name']).to eq(site_name)
 			expect(site_create[:code]).to eq(201).or(200)
+		end
+		it 'deletes a valid site' do
+			sample_site_id = @sites.sample['id']
+			site_delete = @samanage.delete_site(id: sample_site_id)
+
+			expect(site_delete[:code]).to eq(200).or(201)
+		end
+		it 'fails to delete invalid site' do 
+			invalid_site_id = 0
+			expect{@samanage.delete_site(id: invalid_site_id)}.to raise_error(Samanage::NotFound)
 		end
 	end
 end