+++ spec/samanage_api_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 describe Samanage do
 	describe 'API Controller' do
-		before(:each) do
+		before(:all) do
 			TOKEN ||= ENV['SAMANAGE_TEST_API_TOKEN']
 		end
-	  context 'on creation' do
-	  	it 'Requires Email & Token' do
+	  context 'on instantiation' do
+	  	it 'Requires Valid Token' do
 	  		expect{
 	  			api_controller = Samanage::Api.new(token: "invalid token")
 	  			api_controller.authorize
@@ -40,6 +40,16 @@ describe Samanage do
   			admin_email = admins.sample.gsub('+',"%2B")
   			samanage_admin = api_controller.execute(path: "users.json?email=#{admin_email}")
   			expect(samanage_admin[:data].first['role']['name']).to eq('Administrator')
+  		end
+
+  		it 'sets eu datacenter' do
+  			api_controller = Samanage::Api.new(token: 'token', datacenter: 'eu')
+  			expect(api_controller.base_url).to eq('https://apieu.samanage.com/')
+  		end
+
+  		it 'does not set non/eu datacenter' do
+  			api_controller = Samanage::Api.new(token: 'token', datacenter: 'invalid')
+  			expect(api_controller.base_url).to eq('https://api.samanage.com/')
   		end
 	  end
 	end